moved interface and did  a little bit of client work
