+++ internal/interfaces/iBoardService.go
@@ -1,4 +1,4 @@
-package services
+package interfaces
 
 import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
